moving things around; saving changes now so none are lost
