+++ applications/jupyter-extension/nteract_on_jupyter/app/components/editable-title-overlay.tsx
